ca devient (un) long

+++ Jauge 2/src/Partie2/JaugeNaturel/Jauge.java
@@ -11,7 +11,7 @@ public class Jauge {
 	public static Jauge creerJauge(int min, int max) {
 		return new JaugeDistance(min, max);
 	}
-	public static Jauge creerJauge(int min, int max, int val) {
+	public static Jauge creerJauge(long min, long max, long val) {
 		if (min<0||max<0||val<0)
 			return new JaugeNegatif(min, max, val);
 		else
